Add `Enumerable.Zip` (#8)
This pull-request adds support for the `zip(…)` operator. It accepts
another object which can be turned into an `Iterator` by calling
`Iterator.from(…)`. Each element in the result source will be an array
with two element: `[source[i], other[i]]`.

It generates elements as long as both sources generate elements.

closes #7